Ejercicios y modificaciones hasta la tutoria 2
