+++ music_school_gerard/__manifest__.py
@@ -1,0 +1,19 @@
+{
+    'name': 'Music School',
+    'version': '18.0.0.0',
+    'description': 'Manages a music school with students, teachers and classes',
+    'summary': 'Manages a music school with students, teachers and classes',
+    'author': 'Gerard Campos',
+    'license': 'LGPL-3',
+    'category': 'Music School',
+    'depends': [
+        'base'
+    ],
+    'data':[
+        'security/res_groups.xml',
+        'security/ir.model.access.csv',
+        'views/music_school_instrument_views.xml',
+        'views/music_school_student_views.xml',
+        'views/music_school_menuitems.xml',
+    ]
+}
