+++ music_school_gerard/security/ir.model.access.csv
@@ -1,0 +1,5 @@
+id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
+music_school_gerard.access_music_school_student_user,access_music_school_student_user,music_school_gerard.model_music_school_student,music_school_gerard.group_music_school_user,1,0,0,0
+music_school_gerard.access_music_school_student_manager,access_music_school_student_manager,music_school_gerard.model_music_school_student,music_school_gerard.group_music_school_manager,1,1,1,1
+music_school_gerard.access_music_school_instrument_user,access_music_school_instrument_user,music_school_gerard.model_music_school_instrument,music_school_gerard.group_music_school_user,1,0,0,0
+music_school_gerard.access_music_school_instrument_manager,access_music_school_instrument_manager,music_school_gerard.model_music_school_instrument,music_school_gerard.group_music_school_manager,1,1,1,1
