+++ music_school_gerard/security/res_groups.xml
@@ -1,0 +1,35 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+    <record id="ir_module_music_school" model="ir.module.category">
+        <field name="name">Music School</field>
+    </record>
+    
+    <record id="group_music_school_user" model="res.groups">
+        <field name="name">User</field>
+        <field name="category_id" ref="ir_module_music_school"/>
+        <field name="comment">The user will be able to read records</field>
+    </record>
+
+    <record id="group_music_school_manager" model="res.groups">
+        <field name="name">Manager</field>
+        <field name="category_id" ref="ir_module_music_school"/>
+        <field name="implied_ids" eval="[(4, ref('group_music_school_user'))]"></field>
+        <field name="comment">The manager will be able to create, write and delete records</field>
+    </record>
+
+    <record id="music_school_student_personal_rule" model="ir.rule">
+        <field name="name">Personal students</field>
+        <field name="model_id" ref="model_music_school_student"></field>
+        <field name="domain_force">['|',('user_id','=',user.id),('user_id','=',False)]</field>
+        <field name="groups" eval="[(4, ref('group_music_school_user'))]"></field>
+    </record>
+
+    <record id="music_school_student_all_rule" model="ir.rule">
+        <field name="name">All students</field>
+        <field name="model_id" ref="model_music_school_student"></field>
+        <field name="domain_force">[(1,'=',1)]</field>
+        <field name="groups" eval="[(4, ref('group_music_school_manager'))]"></field>
+    </record>
+
+
+</odoo>
