+++ music_school_gerard/views/music_school_instrument_views.xml
@@ -1,0 +1,62 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+<record id="music_school_instrument_action" model="ir.actions.act_window">
+    <field name="name">instrument</field>
+    <field name="res_model">music.school.instrument</field>
+    <field name="view_mode">list,form</field>
+    <field name="context">{}</field> 
+    <field name="help" type="html">
+    <p class="o_view_nocontent_smiling_face">
+        Create your first instrument
+    </p>
+    </field>
+</record>
+
+<record id="music_school_instrument_list" model="ir.ui.view">
+    <field name="name">music.school.instrument.view.list</field>
+    <field name="model">music.school.instrument</field>
+    <field name="arch" type="xml">
+        <list multi_edit="1" string="instrument">
+            <field name="name"/>
+            <field name="type"/>
+            <field name="description" optional="hide"/>
+            <field name="lastRevisionDate" optional="hide"/>
+        </list>
+    </field>
+</record>
+
+<record id="music_school_instrument_view_form" model="ir.ui.view">
+    <field name="name">music.school.instrument.view.form</field>
+    <field name="model">music.school.instrument</field>
+    <field name="arch" type="xml">
+        <form string="instrument">
+            <header>
+            <button string="Generate date" name="generate_last_revision_date" type="object" class="oe_highlight"/> 
+            </header>
+            <sheet>
+                <div class="oe_title">
+                    <h1>
+                        <field name="name" placeholder="instrument"/>
+                    </h1>
+                </div>
+                <group>
+                    <group>
+                        <field name="type" />
+                    </group>
+                </group>
+                <notebook>
+                    <page string="Description">
+                        <field name="description"/>
+                    </page>
+                    <page string="Observations">
+                        <group>
+                            <field name="lastRevisionDate" readonly="1"/>
+                        </group>
+                    </page>
+                </notebook>
+            </sheet>
+        </form>
+    </field>
+</record>
+
+</odoo>
