+++ music_school_gerard/views/music_school_student_views.xml
@@ -4,15 +4,15 @@
     <field name="name">Students</field>
     <field name="res_model">music.school.student</field>
     <field name="view_mode">list,form</field>
-    <field name="context">{}</field>
+    <field name="context">{}</field> 
     <field name="help" type="html">
     <p class="o_view_nocontent_smiling_face">
         Create your first student
     </p>
     </field>
 </record>
-
-<record id="music_school_student_view_list" model="ir.ui.view">
+    
+<record id="music_school_student_list" model="ir.ui.view">
     <field name="name">music.school.student.view.list</field>
     <field name="model">music.school.student</field>
     <field name="arch" type="xml">
@@ -22,7 +22,7 @@
             <field name="phone" optional="show"/>
             <field name="user_id" optional="show"/>
             <field name="birthdate"/>
-            <field name="age" optional="hide"/> 
+            <field name="age" optional="hide"/>
         </list>
     </field>
 </record>
@@ -33,31 +33,24 @@
     <field name="arch" type="xml">
         <form string="Students">
             <header>
-            <button string="Generate reference" name="generate_reference" type="object" class="btn-primary"/>
+            <button string="Generate reference" name="generate_reference" type="object" class="oe_highlight"/> 
             </header>
             <sheet>
-                <div class="oe_button_box" name="button_box">
-                    <button name="action_view_attendances" type="object" class="oe_stat_button" icon="fa-calendar">
-                        <field name="attendances_count" string="Attendances" widget="statinfo"/>
-                    </button>
-                </div>
                 <div class="oe_title">
                     <h1>
                         <field name="name" placeholder="Student"/>
                     </h1>
                 </div>
                 <group>
-                    <group>
-                        <!-- <field name="partner_id"/> -->
+                    <group> 
                         <field name="reference" readonly="1"/>
-                        <field name="email"/>
-                        <field name="phone"/>
-                        <field name="state_id"/>
+                        <field name="email" />
+                        <field name="phone" />
                     </group>
                     <group>
-                        <field name="user_id"/>
-                        <field name="birthdate"/>
-                        <field name="age"/>
+                        <field name="user_id" />
+                        <field name="birthdate" />
+                        <field name="age" />
                     </group>
                 </group>
                 <notebook>
@@ -69,6 +62,4 @@
         </form>
     </field>
 </record>
-
-    
 </odoo>
