--- music_school_nacho/__init__.py
@@ -1,2 +1,0 @@
-from . import models
-from . import wizard
