--- music_school_nacho/data/ir_cron.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-
-    <record id="music_school_exam_cron" model="ir.cron">
-        <field name="name">Finish exams</field>
-        <field name="model_id" ref="model_music_school_exam"/>
-        <field name="interval_number">1</field>
-        <field name="interval_type">days</field>
-        <field name="state">code</field>
-        <field name="code">model.finish_exams()</field>
-    </record>
-
-    <record id="music_school_course_cron" model="ir.cron">
-        <field name="name">Finish courses</field>
-        <field name="model_id" ref="model_music_school_course"/>
-        <field name="interval_number">1</field>
-        <field name="interval_type">days</field>
-        <field name="state">code</field>
-        <field name="code">model.finish_courses()</field>
-    </record>
-
-</odoo>
