--- music_school_nacho/data/ir_sequence.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-
-    <record id="seq_music_school_course" model="ir.sequence">
-        <field name="name">Music School Course</field>
-        <field name="code">music.school.course</field>
-        <field name="prefix">MSC/%(year)s/</field>
-        <field name="padding">5</field>
-    </record>
-
-</odoo>
