--- music_school_nacho/models/__init__.py
@@ -1,10 +1,0 @@
-from . import music_school_student
-from . import music_school_instrument
-from . import music_school_teacher
-from . import music_school_course
-from . import music_school_lesson
-from . import music_school_lesson_attendance
-from . import music_school_room
-from . import music_school_instrument_family
-from . import music_school_exam
-from . import music_school_exam_result
