--- music_school_nacho/models/music_school_instrument_family.py
@@ -1,9 +1,0 @@
-from odoo import models, fields
-
-
-class MusicSchoolInstrumentFamily(models.Model):
-    _name = 'music.school.instrument.family'
-    _description = 'Instruments'
-
-    name = fields.Char(string="Name", required=True, translate=True)
-
