--- music_school_nacho/models/music_school_room.py
@@ -1,19 +1,0 @@
-from odoo import models, fields
-
-class MusicSchoolRoom(models.Model):
-    _name = 'music.school.room'
-    _description = 'Rooms'
-
-    name = fields.Char(string="Name", required=True)
-    capacity = fields.Integer(
-        string="Capacity",
-        help="Maximum number of students that can fit in the room"
-    )
-    location = fields.Char(
-        string="Location",
-        help="Physical location of the room within the school"
-    )
-
-    _sql_constraints = [
-        ('name_unique', 'UNIQUE(name)', 'The room name must be unique.'),
-    ]
