--- music_school_nacho/report/music_school_course_report.xml
@@ -1,83 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-
-    <template id="report_music_school_course">
-        <t t-call="web.html_container">
-            <t t-foreach="docs" t-as="o">
-                <t t-call="web.external_layout">
-                    <div class="page">
-                    <t t-if="o.env.context.get('button1')">
-                        <h1>Desde el botón 1</h1>
-                    </t>
-                    <t t-if="o.env.context.get('button2')">
-                        <h1>Desde el botón 2</h1>
-                    </t>
-                        <h2 t-field="o.name"/>  
-                        <div class="row">
-                            <div class="col-auto col-4 mw-100 mb-2">
-                                <strong>Professor:</strong>
-                                <span t-field="o.teacher_id"/>
-                            </div>
-                            <div class="col-auto col-4 mw-100 mb-2">
-                                <strong>Start Date:</strong>
-                                <span t-field="o.start_date"/>
-                            </div>
-                            <div class="col-auto col-4 mw-100 mb-2">
-                                <strong>End Date:</strong>
-                                <span t-field="o.end_date"/>
-                            </div>
-                        </div>
-                        <table class="table table-sm o_main_table table-borderless mt-4">
-                            <thead>
-                                <tr>
-                                    <th class="text-start">Student</th>
-                                    <th class="text-start">Email</th>
-                                    <th class="text-start">Phone</th>
-                                </tr>
-                            </thead>
-                            <tbody>
-                                <t t-foreach="o.student_ids" t-as="student">
-                                    <tr>
-                                        <td><span t-field="student.name"/></td>
-                                        <td><span t-field="student.email"/></td>
-                                        <td><span t-field="student.phone"/></td>
-                                    </tr>
-                                
-                                </t>
-                            </tbody>
-                        </table>
-                    </div>
-                </t>
-            </t>
-    </t>
-    </template>
-
-
-    <record id="paperformat_music_school" model="report.paperformat">
-        <field name="name">Music School Paperformat</field>
-        <field name="default" eval="True"/>
-        <field name="format">A4</field>
-        <field name="orientation">Portrait</field>
-        <field name="margin_top">40</field>
-        <field name="margin_bottom">15</field>
-        <field name="margin_left">3</field>
-        <field name="margin_right">3</field>
-        <field name="header_line" eval="False"/>
-        <field name="header_spacing">35</field>
-        <field name="dpi">90</field>
-    </record>
-
-    <record id="action_report_music_school_course" model="ir.actions.report">
-        <field name="name">Course</field>
-        <field name="model">music.school.course</field>
-        <field name="report_type">qweb-pdf</field>
-        <field name="report_name">music_school_nacho.report_music_school_course</field>
-        <field name="report_file">music_school_nacho.report_music_school_course</field>
-        <field name="print_report_name">'%s' % object.name</field>
-        <field name="paperformat_id" ref="paperformat_music_school"/>
-        <field name="binding_model_id" ref="model_music_school_course"/>
-        <field name="binding_type">report</field>
-    </record>
-
-
-</odoo>
