--- music_school_nacho/report/music_school_course_simple_report.xml
@@ -1,44 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<odoo>
-
-    <template id="report_music_school_course_simple">
-        <t t-call="web.html_container">
-                <t t-call="web.external_layout">
-                    <div class="page">
-                        <table class="table table-sm o_main_table table-borderless mt-4">
-                            <thead>
-                                <tr>
-                                    <th class="text-start">Course</th>
-                                    <th class="text-start">Date</th>
-                                </tr>
-                            </thead>
-                            <tbody>
-                                <t t-foreach="docs" t-as="o">
-                                    <tr>
-                                        <td><span t-field="o.name"/></td>
-                                        <td><span t-field="o.start_date"/></td>
-                                    </tr>
-                                
-                                </t>
-                            </tbody>
-                        </table>
-                    </div>
-            </t>
-    </t>
-    </template>
-
-
-    <record id="action_report_music_school_course_simple" model="ir.actions.report">
-        <field name="name">Course Simplified</field>
-        <field name="model">music.school.course</field>
-        <field name="report_type">qweb-pdf</field>
-        <field name="report_name">music_school_nacho.report_music_school_course_simple</field>
-        <field name="report_file">music_school_nacho.report_music_school_course_simple</field>
-        <field name="print_report_name">'%s' % object.name</field>
-        <field name="binding_model_id" ref="model_music_school_course"/>
-        <field name="binding_type">report</field>
-    </record>
-
-
-</odoo>
