--- music_school_nacho/report/music_school_exam_report.xml
@@ -1,67 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-
-    <template id="report_music_school_exam">
-        <t t-call="web.html_container">
-            <t t-foreach="docs" t-as="o">
-                <t t-call="web.external_layout">
-                    <div class="page">
-                   
-                        <h2 t-field="o.name"/>  
-                        <div class="row">
-                            <div t-if="o.teacher_id" class="col-auto col-4 mw-100 mb-2">
-                                <strong>Professor:</strong>
-                                <span t-field="o.teacher_id.name"/>
-                            </div>
-                            <div t-if="o.date" class="col-auto col-4 mw-100 mb-2">
-                                <strong>Date:</strong>
-                                <span t-field="o.date"/>
-                            </div>
-                            <div t-if="o.instrument_id" class="col-auto col-4 mw-100 mb-2">
-                                <strong>Instrument:</strong>
-                                <span t-field="o.instrument_id.name"/>
-                            </div>
-                        </div>
-                        <table class="table table-sm o_main_table table-borderless mt-4">
-                            <thead>
-                                <tr>
-                                    <th class="text-start">Student</th>
-                                    <th class="text-start">Score</th>
-                                    <th class="text-end">Passed</th>
-                                </tr>
-                            </thead>
-                            <tbody>
-                                <t t-foreach="o.result_ids" t-as="result">
-                                    <tr>
-                                        <td><span t-field="result.student_id.name"/></td>
-                                        <td><span t-field="result.score"/></td>
-                                        <t t-if="result.passed">
-                                            <td class="text-end">Yes</td>
-                                        </t>
-                                        <t t-else="">
-                                            <td class="text-end">No</td>
-                                        </t>
-                                    </tr>
-                                
-                                </t>
-                            </tbody>
-                        </table>
-                    </div>
-                </t>
-            </t>
-    </t>
-    </template>
-
-
-    <record id="action_report_music_school_exam" model="ir.actions.report">
-        <field name="name">Exam Results</field>
-        <field name="model">music.school.exam</field>
-        <field name="report_type">qweb-pdf</field>
-        <field name="report_name">music_school_nacho.report_music_school_exam</field>
-        <field name="report_file">music_school_nacho.report_music_school_exam</field>
-        <field name="print_report_name">'%s' % object.name</field>
-        <field name="binding_model_id" ref="model_music_school_exam"/>
-        <field name="binding_type">report</field>
-    </record>
-
-</odoo>
