--- music_school_nacho/report/music_school_lesson_report.xml
@@ -1,64 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <template id="report_music_school_lesson">
-        <t t-call="web.html_container">
-            <t t-foreach="docs" t-as="o">
-                <t t-call="web.external_layout">
-                    <div class="page">
-    
-                        <h2 t-field="o.course_id.name"/>  
-                        <div class="row">
-                            <div class="col-auto col-4 mw-100 mb-2">
-                                <strong>Professor:</strong>
-                                <span t-field="o.teacher_id"/>
-                            </div>
-                            <div class="col-auto col-4 mw-100 mb-2">
-                                <strong>Room:</strong>
-                                <span t-field="o.room_id.name"/>
-                            </div>
-                            <div class="col-auto col-4 mw-100 mb-2">
-                                <strong>Date:</strong>
-                                <span t-field="o.date"/>
-                            </div>
-                        </div>
-                        <table class="table table-sm o_main_table table-borderless mt-4">
-                            <thead>
-                                <tr>
-                                    <th class="text-start">Student</th>
-                                    <th class="text-end">Is present</th>
-                                </tr>
-                            </thead>
-                            <tbody>
-                                <t t-foreach="o.attendance_ids" t-as="attendance">
-                                    <tr>
-                                        <td><span t-field="attendance.student_id.name"/></td>
-                                        <t t-if="attendance.is_present">
-                                            <td class="text-end">Yes</td>
-                                        </t>
-                                        <t t-else="">
-                                            <td class="text-end">No</td>
-                                        </t>
-                                    </tr>
-                                
-                                </t>
-                            </tbody>
-                        </table>
-                    </div>
-                </t>
-            </t>
-    </t>
-    </template>
-
-
-    <record id="action_report_music_school_lesson" model="ir.actions.report">
-        <field name="name">Lesson Attendances</field>
-        <field name="model">music.school.lesson</field>
-        <field name="report_type">qweb-pdf</field>
-        <field name="report_name">music_school_nacho.report_music_school_lesson</field>
-        <field name="report_file">music_school_nacho.report_music_school_lesson</field>
-        <field name="print_report_name">'%s' % object.course_id.name</field>
-        <field name="binding_model_id" ref="model_music_school_lesson"/>
-        <field name="binding_type">report</field>
-    </record>
-
-</odoo>
