--- music_school_nacho/security/ir.model.access.csv
@@ -1,19 +1,0 @@
-id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
-music_school_nacho.access_music_school_student_user,access_music_school_student_user,music_school_nacho.model_music_school_student,group_music_school_user,1,0,0,0
-music_school_nacho.access_music_school_student_manager,access_music_school_student_manager,music_school_nacho.model_music_school_student,group_music_school_manager,1,1,1,1
-music_school_nacho.access_music_school_instrument_user,access_music_school_instrument_user,music_school_nacho.model_music_school_instrument,group_music_school_user,1,1,0,0
-music_school_nacho.access_music_school_instrument_manager,access_music_school_instrument_manager,music_school_nacho.model_music_school_instrument,group_music_school_manager,1,1,1,1
-music_school_nacho.access_music_school_teacher_user,access_music_school_teacher_user,music_school_nacho.model_music_school_teacher,group_music_school_user,1,0,0,0
-music_school_nacho.access_music_school_course_user,access_music_school_course_user,music_school_nacho.model_music_school_course,group_music_school_user,1,0,0,0
-music_school_nacho.access_music_school_lesson_user,access_music_school_lesson_user,music_school_nacho.model_music_school_lesson,group_music_school_user,1,0,0,0
-music_school_nacho.access_music_school_room_user,access_music_school_room_user,music_school_nacho.model_music_school_room,group_music_school_user,1,0,0,0
-music_school_nacho.access_music_school_teacher_manager,access_music_school_teacher_manager,music_school_nacho.model_music_school_teacher,group_music_school_manager,1,1,1,1
-music_school_nacho.access_music_school_course_manager,access_music_school_course_manager,music_school_nacho.model_music_school_course,group_music_school_manager,1,1,1,1
-music_school_nacho.access_music_school_lesson_manager,access_music_school_lesson_manager,music_school_nacho.model_music_school_lesson,group_music_school_manager,1,1,1,1
-music_school_nacho.access_music_school_room_manager,access_music_school_room_manager,music_school_nacho.model_music_school_room,group_music_school_manager,1,1,1,1
-music_school_nacho.access_music_school_instrument_family,access_music_school_instrument_family,music_school_nacho.model_music_school_instrument_family,base.group_user,1,1,1,1
-music_school_nacho.access_music_school_lesson_attendance,access_music_school_lesson_attendance,music_school_nacho.model_music_school_lesson_attendance,base.group_user,1,1,1,1
-music_school_nacho.access_music_school_exam,access_music_school_exam,music_school_nacho.model_music_school_exam,base.group_user,1,1,1,1
-music_school_nacho.access_music_school_exam_result,access_music_school_exam_result,music_school_nacho.model_music_school_exam_result,base.group_user,1,1,1,1
-music_school_nacho.access_music_school_course_change_state,access_music_school_course_change_state,music_school_nacho.model_music_school_course_change_state,base.group_user,1,1,1,1
-music_school_nacho.access_music_school_lesson_batch,access_music_school_lesson_batch,music_school_nacho.model_music_school_lesson_batch,base.group_user,1,1,1,1
