--- music_school_nacho/tests/__init__.py
@@ -1,1 +1,0 @@
-from . import test_music_school_course
