--- music_school_nacho/views/music_school_exam_views.xml
@@ -1,143 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-
-    <record id="music_school_exam_action" model="ir.actions.act_window">
-        <field name="name">Exams</field>
-        <field name="res_model">music.school.exam</field>
-        <field name="view_mode">kanban,list,form</field>
-        <field name="domain">[]</field>
-        <field name="context">{'search_default_not_finished': 1}</field>
-        <field name="help" type="html">
-            <p class="o_view_nocontent_smiling_face">
-                Create your first exam
-            </p>
-        </field>
-    </record>
-
-    <!-- <record id="music_school_exam_self_action" model="ir.actions.act_window">
-        <field name="name">Exams</field>
-        <field name="res_model">music.school.exam</field>
-        <field name="view_mode">list,form</field>
-        <field name="domain">[('course_id', '=', active_id)]</field>
-        <field name="context">{'default_course_id': active_id}</field>
-        <field name="help" type="html">
-            <p class="o_view_nocontent_smiling_face">
-                Create your first exam
-            </p>
-        </field>
-    </record> -->
-
-    <record id="music_school_exam_view_tree" model="ir.ui.view">
-        <field name="name">music.school.exam.view.tree</field>
-        <field name="model">music.school.exam</field>
-        <field name="arch" type="xml">
-            <list string="Exams">
-                <field name="name"/>
-                <field name="course_id" optional="show"/>
-                <field name="date" optional="show" />
-                <field name="teacher_id" optional="show"/>
-                <field name="state"/>
-                
-            </list>
-        </field>
-    </record>
-
-    <record id="music_school_exam_view_form" model="ir.ui.view">
-        <field name="name">music.school.exam.view.form</field>
-        <field name="model">music.school.exam</field>
-        <field name="arch" type="xml">
-            <form string="Exam">
-                <header>
-                    <field name="state" widget="statusbar"/>
-                    <button string="Assign students" groups="music_school_nacho.group_music_school_manager" name="assign_students" type="object" class="oe_highlight"/>
-                </header>
-                <sheet>
-                <widget name="web_ribbon" title="Finished"
-                 bg_color="text-bg-info"
-                 invisible="state != 'finished'"/>
-                       
-                <div class="oe_title">
-                    <h1>
-                        <field name="name" placeholder="E.g Flute exam" class="oe_inline"/>
-                    </h1>
-                </div>
-                <group>
-                    <group>
-                        <field name="course_id"/>
-                        <field name="date"/>
-                        <field name="teacher_id"/>
-                    </group>
-
-                    <group>
-                        <field name="instrument_id"/>
-                        <field name="min_score"/>
-                        <field name="max_score"/>
-                    </group>
-                </group>
-                <notebook>
-                    <page name="results" string="Results">
-                        <field name="result_ids">
-                            <list editable="bottom" string="Results">
-                                <field name="student_id" domain="[('id','in', student_ids)]"/>
-                                <field name="score"/>
-                                <field name="passed" widget="boolean_toggle"/>
-                                <field name="notes" optional="hide"/>
-                            </list>
-                        </field>
-                        
-                    </page>
-                </notebook>
-                </sheet>
-            </form>
-        </field>
-    </record>
-
-    <record id="music_school_exam_view_kanban" model="ir.ui.view">
-        <field name="name">music.school.exam.view.kanban</field>
-        <field name="model">music.school.exam</field>
-        <field name="arch" type="xml">
-            <kanban highlight_color="color" default_group_by="state" class="o_kanban_small_column" on_create="quick_create">
-            <field name="name"/>
-            <field name="teacher_id"/>
-            <field name="instrument_id"/>
-            <field name="course_id"/>
-            <field name="date"/>
-                <templates>
-                    <t t-name="menu">
-                        <t t-if="widget.editable"><a role="menuitem" type="open" class="dropdown-item">Edit</a></t>
-                        <t t-if="widget.deletable"><a role="menuitem" type="delete" class="dropdown-item">Delete</a></t>
-                        <field name="color" widget="kanban_color_picker"/> 
-                    </t>
-                    <t t-name="card">
-                        <field name="name" class="fw-bold fs-5"/>
-                        <field name="course_id"/>
-                        <field name="instrument_id" class="text-muted"/>
-                        <field name="date" class="text-muted"/>
-                        <footer>
-                            
-                            <field name="teacher_id" widget="many2one_avatar_user" class="ms-auto"/>
-                           
-                        </footer>
-                    </t>
-                </templates>
-            </kanban>
-        </field>
-    </record>
-
-    <record id="music_school_exam_view_search" model="ir.ui.view">
-        <field name="name">music.school.exam.view.search</field>
-        <field name="model">music.school.exam</field>
-        <field name="arch" type="xml">
-            <search string="Exams">
-                <field name="name" filter_domain="['|',('course_id', 'ilike', self),('name','ilike',self)]"/>
-                <field name="teacher_id"/>
-                <field name="instrument_id"/>
-                <field name="course_id"/>
-                <filter name="exams_done" string="Finished" domain="[('state', '=', 'finished')]" />
-                <filter name="not_finished" string="Not finished" domain="[('state', '!=', 'finished')]" />
-                <filter name="group_by_teacher" string="Teacher" context="{'group_by': 'teacher_id'}" />
-            </search>
-        </field>
-    </record>
-
-</odoo>
