--- music_school_nacho/views/music_school_instrument_views.xml
@@ -1,65 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-
-    <record id="music_school_instrument_action" model="ir.actions.act_window">
-        <field name="name">Instrument</field>
-        <field name="res_model">music.school.instrument</field>
-        <field name="view_mode">list,form</field>
-        <field name="context">{}</field>
-        <field name="help" type="html">
-            <p class="o_view_nocontent_smiling_face">
-            Create your first instrument
-             </p>
-        </field>
-    </record>
-
-    <record id="music_school_instrument_view_list" model="ir.ui.view">
-        <field name="name">music.school.instrument.view.list</field>
-        <field name="model">music.school.instrument</field>
-        <field name="arch" type="xml">
-            <list string="Instruments">
-                <field name="name"/>
-                <field name="family_id" optional="show"/>
-                <!-- <field name="family" optional="show"/> -->
-                <field name="description" optional="hide"/>
-                <field name="is_repaired" widget="boolean_toggle" optional="show"/>
-                <field name="last_maintenance_date" optional="hide"/>
-            </list>
-        </field>
-    </record>
-
-    <record id="music_school_instrument_view_form" model="ir.ui.view">
-        <field name="name">music.school.instrument.view.form</field>
-        <field name="model">music.school.instrument</field>
-        <field name="arch" type="xml">
-            <form string="Instrument">
-            <header>
-            <button string="Instrument maintenance" name="instrument_maintenance" type="object" class="btn-warning"/>
-            </header>
-                <sheet>
-                    <div class="oe_title">
-                        <h1>
-                            <field name="name" placeholder="Instrument"/>
-                        </h1>
-                    </div>
-                    <group>
-                        <field name="family_id"/>
-                        <!-- <field name="family" widget="radio"/> -->
-                    </group>
-                    <notebook>
-                    <page string="Description">
-                        <field name="description"/>
-                    </page>
-                    <page string="Observations">
-                    <group>
-                        <field name="last_maintenance_date" readonly="1"/>
-                        <field name="is_repaired" widget="boolean_toggle"/>
-                    </group>
-                    </page>
-                </notebook>
-                </sheet>
-            </form>
-        </field>
-    </record>
-
-</odoo>
