--- music_school_nacho/views/music_school_lesson_attendance_views.xml
@@ -1,46 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-
-    <record id="music_school_lesson_attendance_action" model="ir.actions.act_window">
-        <field name="name">Attendance</field>
-        <field name="res_model">music.school.lesson.attendance</field>
-        <field name="view_mode">list,form</field>
-        <field name="domain">[]</field>
-        <field name="context">{}</field>
-        <field name="help" type="html">
-            <p class="o_view_nocontent_smiling_face">
-                Create your first attendance record
-            </p>
-        </field>
-    </record>
-    <record id="music_school_lesson_attendance_view_tree" model="ir.ui.view">
-        <field name="name">music.school.lesson.attendance.view.tree</field>
-        <field name="model">music.school.lesson.attendance</field>
-        <field name="arch" type="xml">
-            <list string="Attendance">
-                <field name="student_id"/>
-                <field name="lesson_id"/>
-                <field name="is_present"/>
-                <field name="notes" optional="hide"/>
-                <field name="date" optional="show"/>
-            </list>
-        </field>
-    </record>
-
-    <record id="music_school_lesson_attendance_simple_view_tree" model="ir.ui.view">
-        <field name="name">music.school.lesson.attendance.simple.view.tree</field>
-        <field name="model">music.school.lesson.attendance</field>
-        <field name="priority" eval="99"/>
-        <field name="arch" type="xml">
-            <list string="Attendance" editable="bottom">
-                <field name="sequence" widget="handle"/>
-                <field name="student_id"/>
-                <field name="lesson_id"/>
-                <field name="is_present" widget="boolean_toggle"/>
-                <field name="date" optional="show"/>
-                <field name="notes" optional="hide"/>
-            </list>
-        </field>
-    </record>
-
-</odoo>
