--- music_school_nacho/views/music_school_lesson_views.xml
@@ -1,159 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-
-    <record id="music_school_lesson_action" model="ir.actions.act_window">
-        <field name="name">Lesson</field>
-        <field name="res_model">music.school.lesson</field>
-        <field name="view_mode">list,kanban,pivot,graph,calendar,form</field>
-        <field name="domain">[]</field>
-        <field name="context">{}</field>
-        <field name="help" type="html">
-            <p class="o_view_nocontent_smiling_face">
-            Create your first lesson
-             </p>
-        </field>
-    </record>
-
-    <record id="music_school_lesson_view_tree" model="ir.ui.view">
-        <field name="name">music.school.lesson.view.tree</field>
-        <field name="model">music.school.lesson</field>
-        <field name="arch" type="xml">
-            <list string="Lesson">
-                <field name="date"/>
-                <field name="teacher_id"/>
-                <field name="course_id" optional="show"/>
-                <field name="room_id" optional="show"/>
-                <field name="duration" widget="float_time" optional="hide"/>
-                <field name="notes" optional="hide"/>
-                <field name="state" optional="show"/>
-            </list>
-        </field>
-    </record>
-
-    <record id="music_school_lesson_view_form" model="ir.ui.view">
-        <field name="name">music.school.lesson.view.form</field>
-        <field name="model">music.school.lesson</field>
-        <field name="arch" type="xml">
-            <form string="lesson">
-            <header>
-            <field name="state" widget="statusbar"/>
-            </header>
-                <sheet>
-                <div class="oe_title">
-                    <h1>
-                        <field name="course_id" placeholder="Course"/>
-                    </h1>
-                </div>
-                    <group>
-                    <group>
-                        <field name="teacher_id"/>
-                        <field name="room_id"/>
-                    </group>
-                    <group>
-                        <field name="date"/>
-                        <field name="duration" widget="float_time"/>
-                    </group>
-                        
-                    </group>
-                    <notebook>
-                        <page name="attendances" string="Attendances">
-                            <field name="attendance_ids" context="{'default_is_present':True,'default_date': date,'list_view_ref': 'music_school_nacho.music_school_lesson_attendance_simple_view_tree'}"/>
-                        </page>
-                        <page name="notes" string="Notes">
-                            <field name="notes"/>
-                        </page>
-                    </notebook>
-                </sheet>
-                <chatter/>
-            </form>
-        </field>
-    </record>
-
-    <record id="music_school_lesson_view_search" model="ir.ui.view">
-        <field name="name">music.school.lesson.view.search</field>
-        <field name="model">music.school.lesson</field>
-        <field name="arch" type="xml">
-            <search string="Lesson">
-                <field name="course_id" string="Course"/>
-                <field name="teacher_id" string="Teacher"/>
-                <field name="room_id" string="Room"/>
-                <filter name="not_assigned" string="Not Assigned" domain="[('teacher_id', '=', False)]"/>
-                <filter name="date" string="Date" context="{'group_by': 'date'}" />
-                <filter name="room" string="Room" context="{'group_by': 'room_id'}" />
-                <filter name="teacher" string="Teacher" context="{'group_by': 'teacher_id'}" />
-                <filter name="filter_yesterday" string="Ayer"
-                    domain="[('date', '&gt;=', (context_today() - relativedelta(days=1)).strftime('%%Y-%%m-%%d 00:00:00')),
-                         ('date', '&lt;', context_today().strftime('%%Y-%%m-%%d 00:00:00'))]"/>
-
-            </search>
-        </field>
-    </record>
-
-    <record id="music_school_lesson_view_kanban" model="ir.ui.view">
-        <field name="name">music.school.lesson.view.kanban</field>
-        <field name="model">music.school.lesson</field>
-        <field name="arch" type="xml">
-            <kanban highlight_color="color" default_group_by="state" class="o_kanban_small_column" on_create="quick_create">
-            <field name="course_id"/>   
-            <field name="teacher_id"/>
-            <field name="room_id"/>
-            <field name="date"/>
-                <templates>
-                    <t t-name="menu">
-                        <t t-if="widget.editable"><a role="menuitem" type="open" class="dropdown-item">Edit</a></t>
-                        <t t-if="widget.deletable"><a role="menuitem" type="delete" class="dropdown-item">Delete</a></t>
-                        <field name="color" widget="kanban_color_picker"/> 
-                    </t>
-                    <t t-name="card">
-                        <field name="course_id" class="fw-bold fs-5"/>
-                        <field name="date"/>
-                        <footer>
-                            <div class="d-flex mt-auto align-items-center">
-                                <field name="room_id" class="me-2"/>
-                            </div>
-                            <field name="teacher_id" widget="many2one_avatar_user" class="ms-auto"/>
-                           
-                        </footer>
-                    </t>
-                </templates>
-            </kanban>
-        </field>
-    </record>
-
-    <record id="music_school_lesson_view_pivot" model="ir.ui.view">
-        <field name="name">music.school.lesson.view.pivot</field>
-        <field name="model">music.school.lesson</field>
-        <field name="arch" type="xml">
-            <pivot string="Lesson">
-                <field name="date" type="row"/>
-                <field name="course_id" type="col"/>
-                <field name="duration" widget="float_time" type="measure"/>
-            </pivot>
-        </field>
-    </record>
-
-    <record id="music_school_lesson_view_calendar" model="ir.ui.view">
-        <field name="name">music.school.lesson.view.calendar</field>
-        <field name="model">music.school.lesson</field>
-        <field name="arch" type="xml">
-            <calendar string="Lesson" date_start="date" color="course_id">
-                <field name="course_id"/>
-                <field name="teacher_id"/>
-                <field name="room_id"/>
-
-            </calendar>
-        </field>
-    </record>
-
-    <record id="music_school_lesson_view_graph" model="ir.ui.view">
-        <field name="name">music.school.lesson.view.graph</field>
-        <field name="model">music.school.lesson</field>
-        <field name="arch" type="xml">
-            <graph string="Lesson" type="bar">
-                <field name="course_id"/>
-                <field name="duration" widget="float_time" type="measure"/>
-            </graph>
-        </field>
-    </record>
-
-</odoo>
