--- music_school_nacho/views/music_school_menuitems.xml
@@ -1,37 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-
-<!-- <menuitem
-    id="music_school_menu"
-    name="Music School"
-    sequence="10"/>
-
-<menuitem
-    id="music_school_student_menu"
-    name="Student"
-    action="music_school_student_action"
-    parent="music_school_menu"
-    sequence="10"/> -->
-
-<menuitem id="music_school_menu" name="Music School" sequence="10" web_icon="music_school_nacho,static/img/logo.png">
-    <menuitem id="music_school_members" name="Members" sequence="10">
-        <menuitem id="music_school_student_menu" name="Student" action="music_school_student_action" sequence="10"/>
-        <menuitem id="music_school_teacher_menu" name="Teacher" action="music_school_teacher_action" sequence="20"/>
-    </menuitem>
-    <menuitem id="music_school_organization" name="Organization" sequence="20">
-        <menuitem id="music_school_course_menu" name="Course" action="music_school_course_action" sequence="10"/>
-        <menuitem id="music_school_room_menu" name="Room" action="music_school_room_action" sequence="20"/>
-        <menuitem id="music_school_lesson_menu" name="Lesson" action="music_school_lesson_action" sequence="30"/>
-    </menuitem>
-    <menuitem id="music_school_instruments" name="Instruments" sequence="30">
-        <menuitem id="music_school_instrument_menu" name="Instrument" action="music_school_instrument_action" sequence="10"/>
-    </menuitem>
-    <menuitem id="music_school_evaluation" name="Evaluation" sequence="40">
-        <menuitem id="music_school_exam_menu" name="Exams" action="music_school_exam_action" sequence="10"/>
-    </menuitem>
-    <menuitem id="music_school_attendance" name="Attendance" sequence="50">
-        <menuitem id="music_school_lesson_attendance_menu" name="Attendance" action="music_school_lesson_attendance_action" sequence="10"/>
-    </menuitem>
-</menuitem>
-
-</odoo>
