--- music_school_nacho/views/music_school_room_views.xml
@@ -1,49 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-
-    <record id="music_school_room_action" model="ir.actions.act_window">
-        <field name="name">Room</field>
-        <field name="res_model">music.school.room</field>
-        <field name="view_mode">list,form</field>
-        <field name="domain">[]</field>
-        <field name="context">{}</field>
-        <field name="help" type="html">
-            <p class="o_view_nocontent_smiling_face">
-            Create your first room
-             </p>
-        </field>
-    </record>
-
-    <record id="music_school_room_view_tree" model="ir.ui.view">
-        <field name="name">music.school.room.view.tree</field>
-        <field name="model">music.school.room</field>
-        <field name="arch" type="xml">
-            <list string="Room">
-                <field name="name"/>
-                <field name="capacity" optional="show"/>
-                <field name="location" optional="show"/>
-            </list>
-        </field>
-    </record>
-
-    <record id="music_school_room_view_form" model="ir.ui.view">
-        <field name="name">music.school.room.view.form</field>
-        <field name="model">music.school.room</field>
-        <field name="arch" type="xml">
-            <form string="room">
-                <sheet>
-                <div class="oe_title">
-                    <h1>
-                        <field name="name" placeholder="Room"/>
-                    </h1>
-                </div>
-                    <group>
-                        <field name="capacity"/>
-                        <field name="location"/>
-                    </group>
-                </sheet>
-            </form>
-        </field>
-    </record>
-
-</odoo>
