--- music_school_nacho/views/music_school_teacher_views.xml
@@ -1,56 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-
-    <record id="music_school_teacher_action" model="ir.actions.act_window">
-        <field name="name">Teacher</field>
-        <field name="res_model">music.school.teacher</field>
-        <field name="view_mode">list,form</field>
-        <field name="domain">[]</field>
-        <field name="context">{}</field>
-        <field name="help" type="html">
-            <p class="o_view_nocontent_smiling_face">
-            Create your first teacher
-             </p>
-        </field>
-    </record>
-
-    <record id="music_school_teacher_view_tree" model="ir.ui.view">
-        <field name="name">music.school.teacher.view.tree</field>
-        <field name="model">music.school.teacher</field>
-        <field name="arch" type="xml">
-            <list string="">
-                <field name="name"/>
-                <field name="email" optional="show"/>
-                <field name="phone" optional="show"/>
-                <field name="level" optional="show"/>
-            </list>
-        </field>
-    </record>
-
-    <record id="music_school_teacher_view_form" model="ir.ui.view">
-        <field name="name">music.school.teacher.view.form</field>
-        <field name="model">music.school.teacher</field>
-        <field name="arch" type="xml">
-            <form string="Teacher">
-                <sheet>
-                <div class="oe_button_box" name="button_box">
-                    <button invisible="course_count == 0" name="action_view_courses" type="object" class="oe_stat_button" icon="fa-book">
-                        <field name="course_count" string="Courses" widget="statinfo"/>
-                    </button>
-                </div>
-                <div class="oe_title">
-                    <h1>
-                        <field name="name" placeholder="Teacher"/>
-                    </h1>
-                </div>
-                    <group>
-                        <field name="level"/>
-                        <field name="email"/>
-                        <field name="phone"/>
-                    </group>
-                </sheet>
-            </form>
-        </field>
-    </record>
-
-</odoo>
