--- music_school_nacho/wizard/__init__.py
@@ -1,2 +1,0 @@
-from . import music_school_course_change_state
-from . import music_school_lesson_batch
