--- music_school_nacho/wizard/music_school_course_change_state.xml
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-
-    <record id="music_school_course_change_state_action" model="ir.actions.act_window">
-        <field name="name">Change State</field>
-        <field name="res_model">music.school.course.change.state</field>
-        <field name="view_mode">form</field>
-        <field name="target">new</field>
-        <field name="binding_model_id" ref="music_school_nacho.model_music_school_course"/>
-        <field name="binding_view_types">list</field>
-    </record>
-
-    <record id="music_school_course_change_state_view_form" model="ir.ui.view">
-        <field name="name">music.school.course.change.state.view.form</field>
-        <field name="model">music.school.course.change.state</field>
-        <field name="arch" type="xml">
-            <form string="Change Course State">
-                <group>
-                    <field name="state"/>
-                </group>
-                <footer>
-                    <button string="Change State" type="object" name="action_change_state" class="btn-primary"/>
-                    <button string="Cancel" class="btn-secondary" special="cancel"/>
-                </footer>
-            </form>
-        </field>
-    </record>
-</odoo>
