--- music_school_nacho_sale/__manifest__.py
@@ -1,18 +1,0 @@
-{
-    'name': 'Music School Nacho Sale',
-    'version': '18.0.0.0.0',
-    'description': 'Management sales for a music school.',
-    'summary': 'Management sales for a music school.',
-    'author': 'Nacho Serra',
-    'license': 'LGPL-3',
-    'category': 'sale',
-    'depends': [
-        'sale','music_school_nacho'
-    ],
-    'data': [
-        'views/music_school_course.xml',
-        'views/sale_order_views.xml'
-    ],
-    'application': False,
-    
-}
