--- music_school_nacho_sale/models/__init__.py
@@ -1,2 +1,0 @@
-from . import music_school_course
-from . import sale_order
