--- music_school_nacho_sale/views/sale_order_views.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-
-    <record id="sale_order_view_form" model="ir.ui.view">
-        <field name="name">sale.order.view.form</field>
-        <field name="inherit_id" ref="sale.view_order_form"/>
-        <field name="model">sale.order</field>
-        <field name="arch" type="xml">
-            <field name="user_id" position="before">
-                <field name="course_id"/>
-            </field>
-        </field>
-    </record>
-
-</odoo>
